Rename Entity -> BlockEntity, Attribute -> BlockAttribute, Index -> BlockIndex
