+++ src/Zebra/Data.hs
@@ -5,9 +5,7 @@ module Zebra.Data (
 
 import           Zebra.Data.Block as X
 import           Zebra.Data.Encoding as X
-import           Zebra.Data.Entity as X
 import           Zebra.Data.Fact as X
-import           Zebra.Data.Index as X
 import           Zebra.Data.Schema as X
 import           Zebra.Data.Table as X
 import           Zebra.Data.Table.Mutable as X
